# Reduce code duplication in randnt and marginally improve compiler performance.