I noticed that there was a large amount of code duplication in `randnt` (the function called to sample from truncated normal distributions). Furthermore, the meaning of the code seems somewhat unclear when one first looks at it.

I made some changes which I believe make the relationship between the different `if` statements clearer.

I performed some testing and found this version to be around as performant as the previous implementation (possibly marginally faster), while JIT memory allocations have also been marginally reduced in quantity. It seems to me that the compiler (as of Julia v1.5.3) has a slightly easier time understanding the intent of the code for the exponential distribution sampling case of the algorithm.

Current implementation:
```
julia> # trigger exponential distribution case

julia> t = Distributions.truncated(Distributions.Normal(0,1), 10, 20)
Truncated(Main.Distributions.Normal{Float64}(μ=0.0, σ=1.0), range=(10.0, 20.0))

julia> @time rand(t, 1000);
  0.067620 seconds (156.99 k allocations: 7.762 MiB)

julia> @time rand(t, 1000);
  0.000048 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000049 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000049 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000048 seconds (1 allocation: 7.938 KiB)

julia> # trigger uniform distribution case

julia> t2 = Distributions.truncated(Distributions.Normal(0,1), -.001, .001)
Truncated(Main.Distributions.Normal{Float64}(μ=0.0, σ=1.0), range=(-0.001, 0.001))

julia> @time rand(t2, 1000);
  0.000024 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000026 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000027 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000026 seconds (1 allocation: 7.938 KiB)
```

This proposal:
```
julia> # trigger exponential distribution case

julia> t = Distributions.truncated(Distributions.Normal(0,1), 10, 20)
Truncated(Main.Distributions.Normal{Float64}(μ=0.0, σ=1.0), range=(10.0, 20.0))

julia> @time rand(t, 1000);
  0.064036 seconds (150.70 k allocations: 7.506 MiB)

julia> @time rand(t, 1000);
  0.000049 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000046 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000046 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t, 1000);
  0.000046 seconds (1 allocation: 7.938 KiB)

julia> # trigger uniform distribution case

julia> t2 = Distributions.truncated(Distributions.Normal(0,1), -.001, .001)
Truncated(Main.Distributions.Normal{Float64}(μ=0.0, σ=1.0), range=(-0.001, 0.001))

julia> @time rand(t2, 1000);
  0.000024 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000024 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000025 seconds (1 allocation: 7.938 KiB)

julia> @time rand(t2, 1000);
  0.000025 seconds (1 allocation: 7.938 KiB)
```